Separate screens into their corresponding files
