+++ src/screens/Favourites/FavouritesScreen.tsx
@@ -1,0 +1,9 @@
+import { Text, View} from 'react-native';
+
+export const FavouritesScreen= () => {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Favourites!</Text>
+    </View>
+  );
+}
