+++ src/screens/Friends/FriendsScreen.tsx
@@ -1,0 +1,9 @@
+import { Text, View} from 'react-native';
+
+export const FriendsScreen= () => {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Friends!</Text>
+    </View>
+  );
+}
