+++ src/screens/Map/MapScreen.tsx
@@ -1,0 +1,23 @@
+import MapView from 'react-native-maps';
+import {View, StyleSheet, Dimensions } from 'react-native';
+
+export const MapScreen= () => {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <MapView style={styles.map} />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  map: {
+    width: Dimensions.get('window').width,
+    height: Dimensions.get('window').height,
+  },
+});
