+++ src/screens/Profile/ProfileScreen.tsx
@@ -1,0 +1,9 @@
+import { Text, View} from 'react-native';
+
+export const ProfileScreen = () => {
+  return (
+    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>Profile!</Text>
+    </View>
+  );
+}
